requirements.txt
python-dotenv

tiktoken

openai
google-genai